# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Reloop API
  version: 1.0.0
  description: |
    API oficial del backend de Reloop.
    Incluye autenticación JWT, publicaciones, ofertas, carrito,
    compras, favoritos, ratings, usuarios y comentarios.
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #############################
  # GLOBAL SCHEMAS
  #############################
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, example: vale@example.com }
        username: { type: string }
        name: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        country: { type: string, nullable: true }
        birthDate:
          type: string
          format: date
          nullable: true
        avatar:
          type: string
          example: https://example.com/avatar.png
        createdAt:
          type: string
          format: date-time
...
    ListingPhoto:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
    Listing:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        subCategory: { type: string, nullable: true }
        subSubCategory: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        color: { type: string, nullable: true }
        seller:
          $ref: "#/components/schemas/User"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ListingPhoto"
        createdAt:
          type: string
          format: date-time
    ListingCreateRequest:
      type: object
      required: [title, price, condition, category]
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        subCategory: { type: string }
        subSubCategory: { type: string }
        brand: { type: string }
        color: { type: string }
        photos:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
    ListingUpdateRequest:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        subCategory: { type: string }
        subSubCategory: { type: string }
        brand: { type: string }
        color: { type: string }
        photos:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
...
paths:
...
  /listings/all:
    get:
      tags: [Listings]
      summary: Listar publicaciones con filtros, paginación y búsqueda
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: subCategory
          schema: { type: string }
        - in: query
          name: subSubCategory
          schema: { type: string }
        - in: query
          name: condition
          schema: { type: string }
        - in: query
          name: brand
          schema: { type: string }
        - in: query
          name: color
          schema: { type: string }
        - in: query
          name: sizeTop
          schema: { type: string }
        - in: query
          name: sizeBottom
          schema: { type: string }
        - in: query
          name: sizeShoe
          schema: { type: string }
        - in: query
          name: sizeAccessory
          schema: { type: string }
        - in: query
          name: sizeKids
          schema: { type: string }
        - in: query
          name: sizeKidsShoe
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
            default: newest
      responses:
        "200":
          description: Lista paginada de publicaciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /listings/mine:
    get:
      tags: [Listings]
      summary: Listar publicaciones creadas por el usuario autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Publicaciones propias
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                  total:
                    type: integer
        "401":
          description: Token inválido
  /listings/create:
    post:
      tags: [Listings]
      summary: Crear una publicación
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingCreateRequest"
      responses:
        "201":
          description: Publicación creada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          description: Error de validación
        "401":
          description: Token inválido o faltante
  /listings/update/{id}:
    patch:
      tags: [Listings]
      summary: Actualizar una publicación propia
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdateRequest"
      responses:
        "200":
          description: Publicación actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "403":
          description: No autorizado
        "404":
          description: Publicación no encontrada
  /listings/delete/{id}:
    delete:
      tags: [Listings]
      summary: Eliminar una publicación propia
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Publicación eliminada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        "403":
          description: No autorizado
        "404":
          description: Publicación no encontrada
  /listings/detail/{id}:
    get:
      tags: [Listings]
      summary: Obtener detalle completo de una publicación
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalle de la publicación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Publicación no encontrada
  /listings/recommended:
    get:
      tags: [Listings]
      summary: Listado de publicaciones recomendadas
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Publicaciones recomendadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/newest:
    get:
      tags: [Listings]
      summary: Listado aleatorio de las publicaciones más nuevas
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Publicaciones recientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  ###############################################
  # FAVORITES
  ###############################################
...